+++ app/src/main/java/io/familymoments/app/feature/home/viewmodel/HomeViewModel.kt
@@ -7,7 +7,7 @@ import io.familymoments.app.core.network.datasource.UserInfoPreferencesDataSourc
 import io.familymoments.app.core.network.repository.FamilyRepository
 import io.familymoments.app.core.network.repository.PostRepository
 import io.familymoments.app.core.util.DateFormatter
-import io.familymoments.app.core.util.DateFormatter.convertCreatedAtToLocalDate
+import io.familymoments.app.core.util.convertCreatedAtToLocalDate
 import io.familymoments.app.feature.home.uistate.HomeUiState
 import io.familymoments.app.feature.home.uistate.PostPopupType
 import kotlinx.coroutines.flow.MutableStateFlow
